+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 ## 2.0.*
 
+`2022-03-03`
+
+- Add Azure DevOps pipelines system diagnostics support ([TBU](TBU))
+
 `2021-02-15`
 
 - Improve release status retry mechanism ([60]([TBU](https://github.com/dmitryserbin/azdev-release-orchestrator/issues/60)))