+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 3.0.*
 
+`2022-05-26`
+
+- Improve force stage start mechanism update stage ([76](https://github.com/dmitryserbin/azdev-release-orchestrator/pull/76))
+
 `2022-04-30`
 
 - Fixed BuildTags parameter not assigned ([72](https://github.com/dmitryserbin/azdev-release-orchestrator/pull/72))